Fix ambiguous trigonometry, and add the design preview
Two compile fixes and one design change that the preview turned up.

Compile: `cos` and `sin` were ambiguous in CompassRose and CelebrationView.
Swift's implicit CGFloat/Double conversion makes both overloads viable when a
CGFloat and a Double meet in the same expression, so the call cannot be
resolved. Both now do the trigonometry in Double and convert once at the end.

Design: ThemedBackground gains a bottom scrim. Every backdrop runs warm and
bright at its foot, which is exactly where the floating bar sits — the tab
labels were landing light-on-light and effectively disappearing. Deep at the
top, the sunset band through the middle, deep again at the bottom also just
looks better. One layer fixes both the iOS 18 mesh and the iOS 17 fallback.

The preview is a single self-contained page — docs/preview/index.html — with
all nine screens in all six themes, built from template.html by
Tools/build_preview.py. It exists because this app cannot be compiled here, so
the design needed some surface that could actually be looked at. Colours are
copied from Theme.swift and the arrow is the same bezier as ArrowShape, so it
is a faithful mirror rather than an impression. Placeholder photography is
painted by Tools/make_preview_images.py: abstractions, not pictures of real
places.

Verified by rendering all 54 screen/theme combinations in headless Chromium
and reviewing them. That pass also caught four bugs in the mockup itself: the
backdrop gradient ran the wrong way, the theme tiles and rail chips likewise,
a lock-screen widget's distance was painted under its own inner disc, and the
frosted compass plate was covering the tick ring it sits behind.

+++ Shared/DesignSystem/CelebrationView.swift
@@ -29,10 +29,12 @@ struct CelebrationView: View {
                     guard elapsed <= lifetime else { continue }
 
                     let t = elapsed
-                    // Ballistic, with drag so it settles instead of flying off screen.
+                    // Ballistic, with drag so it settles instead of flying off screen. The
+                    // trigonometry stays in Double and converts once, because mixing Double and
+                    // CGFloat in one expression makes cos and sin ambiguous.
                     let drag = 1 - exp(-t * 1.6)
-                    let x = origin.x + cos(angle) * speed * drag
-                    let y = origin.y + sin(angle) * speed * drag + 190 * t * t
+                    let x = origin.x + CGFloat(cos(angle) * speed * drag)
+                    let y = origin.y + CGFloat(sin(angle) * speed * drag + 190 * t * t)
 
                     let progress = t / lifetime
                     let opacity = (1 - progress) * (1 - progress)
